Format meal price and fixed word-wrapping
Format the meal price using the java currency formatter. The prices are now displayed with commas instead of dots and a Euro sign is appended as well.

To prevent word-wrapping of the menu price, the layout should use WRAP_CONTENT instead of a fixed 120px width. To allow the mealLayout to take the remaining space, its weight is set to '1'.